Temporarily removed need to have multi-case catch.
We either need to remove the restriction of having source compatibility
set to 1.6 in Gradle (in the build settings) or not use multicase catch.
For now I've removed the multi-case, but we might want that later, so we
can discuss what to do. Or maybe I'm just doing something wrong, I dunno.