Provided background motivation for the method.
Hopefully clears up the issues A, B of the reviewer since they felt out
of place there. It's beneficial to have the primary contributions in the
introduction like other articles to get an idea of what the primary stuff
our research has found and what differences there are to others. This will
give the proposed method part also a reasonable flow. Perhaps make the last
sentence of the proposed section better a bit later.